Move site from pages route to app route Fixes #49
