+++ app/(totbh)/components/BackToHomeLink.js
@@ -2,7 +2,7 @@
 
 import { usePathname } from "next/navigation";
 import Link from "next/link";
-import styles from "../../../components/layout.module.css";
+import styles from "./backtohome.module.css";
 
 export default function BackToHomeLink() {
   return usePathname() === "/" ? null : (
