+++ app/(totbh)/components/DisplayDate.js
@@ -1,0 +1,6 @@
+import { parseISO, format } from "date-fns";
+
+export default function DisplayDate({ dateString }) {
+  const date = parseISO(dateString);
+  return <time dateTime={dateString}>{format(date, "LLLL d, yyyy")}</time>;
+}
