+++ app/(totbh)/layout.module.css
@@ -9,7 +9,3 @@
   flex-direction: column;
   align-items: center;
 }
-
-.backToHome {
-  margin: 3rem 0 0;
-}
